requirements.txt
streamlit
newspaper3k
scikit-learn
pandas
